+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 ## Current Release 
+### 0.118.0 
+**Release Date:** Thu May 24 21:26:46 UTC 2018     
+## Previous Releases 
+* Feature - aggregated metrics query
+### 0.117.0 
+**Release Date:** Thu May 24 16:28:07 UTC 2018     
+* Fix - Removed 'createdTimestamp' requirement for model validation in Admin Tenant, Tenant User, Tenant Domain, Tenant Ingestion Profile, Tenant Threshold Profile, Tenant Metadata.
+### 0.116.0 
+**Release Date:** Wed May 23 18:25:07 UTC 2018     
+### 0.115.0 
+**Release Date:** Wed May 23 14:32:49 UTC 2018     
+* Fix - remove ThresholdProfileSet from the domain model
+### 0.114.0 
+**Release Date:** Wed May 23 13:14:24 UTC 2018     
 ### 0.113.0 
 **Release Date:** Wed May 16 16:41:53 UTC 2018     
-## Previous Releases 
 * Fix - handle lookup not found errors from druid
 ### 0.112.0 
 **Release Date:** Fri May 11 18:47:21 UTC 2018